bindings: pods uses entities/types
